moving start/stop playmode into its own functions file, try exiting play mode on scene loading
